requirements.txt
# Core dependencies
requests>=2.31.0
flask>=2.3.2
pandas>=2.0.3
numpy>=1.24.3
websocket-client>=1.6.1

# Database
sqlite3  # Built-in with Python

# Monitoring & Testing
psutil>=5.9.0
prometheus_client>=0.18.0

# Advanced Optimization Features
redis>=4.5.0           # High-performance caching
aioredis>=2.0.1         # Async Redis client
asyncio                 # Built-in with Python 3.7+
scikit-learn>=1.3.0     # Machine learning for strategy optimization
joblib>=1.3.0           # Model persistence and parallel processing
ccxt>=4.0.0             # Async cryptocurrency exchange integration
pyyaml>=6.0             # Production configuration management

# Machine Learning Integration
tensorflow>=2.13.0      # Deep learning framework for LSTM models
keras>=2.13.1           # High-level neural networks API
numpy>=1.24.0           # Advanced numerical computing
pandas>=2.0.0           # Enhanced data manipulation
scipy>=1.11.0           # Statistical functions and optimization
ta-lib>=0.4.25          # Technical analysis indicators
matplotlib>=3.7.0       # Plotting and visualization
seaborn>=0.12.0         # Statistical data visualization

# NLP and Sentiment Analysis
transformers>=4.30.0    # BERT models for sentiment analysis (optional)
torch>=2.0.0            # PyTorch backend for transformers (optional)
nltk>=3.8.1             # Natural language processing toolkit (optional)
textblob>=0.17.1        # Simple NLP library (optional)

# Optional for enhanced features
python-binance>=1.0.19  # Official Binance Python SDK
ta-lib>=0.4.25          # Technical Analysis Library (optional)
plotly>=5.15.0          # For backtesting visualizations (optional)

# Development & Testing
pytest>=7.4.0
pytest-mock>=3.11.1
coverage>=7.2.0

# Production Dependencies (optional)
# gunicorn>=21.2.0        # WSGI HTTP Server
# nginx                   # Reverse proxy (system package)
# docker>=6.1.0           # Container runtime
# docker-compose>=2.20.0  # Container orchestration